classifying active fires by land use issue #13

+++ ams_background_tasks/tools/classify_by_land_use.py
@@ -20,7 +20,12 @@ from shapely.geometry import mapping
 from ams_background_tasks.database_utils import DatabaseFacade
 from ams_background_tasks.log import get_logger
 from ams_background_tasks.tools.common import (
+    ACTIVE_FIRES_CLASSNAME,
+    ACTIVE_FIRES_INDICATOR,
+    DETER_INDICATOR,
+    INDICATORS,
     PIXEL_LAND_USE_AREA,
+    RISK_CLASSNAME,
     get_prefix,
     is_valid_biome,
     read_spatial_units,
@@ -53,12 +58,25 @@ logger = get_logger(__name__, sys.stdout)
 )
 @click.option("--biome", type=str, required=True, help="Biome.", multiple=True)
 @click.option(
+    "--indicators",
+    type=str,
+    required=True,
+    default=INDICATORS,
+    multiple=True,
+    help=f"Indicator list ({', '.join(INDICATORS)})",
+)
+@click.option(
     "--drop-tmp", is_flag=True, default=False, help="Drop the temporary tables."
 )
-def main(db_url: str, land_use_dir: str, all_data: bool, biome: str, drop_tmp: bool):
+def main(
+    db_url: str,
+    land_use_dir: str,
+    all_data: bool,
+    biome: tuple,
+    drop_tmp: bool,
+    indicators: tuple,
+):
     """Cross the indicators with the land use image and group them by spatial units."""
-    assert not all_data
-
     db_url = os.getenv("AMS_DB_URL") if not db_url else db_url
     logger.debug(db_url)
     assert db_url
@@ -68,34 +86,34 @@ def main(db_url: str, land_use_dir: str, all_data: bool, biome: str, drop_tmp: b
 
     reset_land_use_tables(db_url=db_url, is_temp=True)
 
-    biome_list = list(biome)
+    assert Path(land_use_dir).exists()
 
-    for index, _biome in enumerate(biome_list):
-        logger.info("processing biome %s", _biome)
-        assert is_valid_biome(biome=_biome)
-
-        land_use_image = Path(land_use_dir) / f"{_biome}_land_use.tif"
-        logger.debug(land_use_image)
-        assert land_use_image.exists()
-
-        process_deter_land_structure(
+    if DETER_INDICATOR in indicators:
+        logger.info("processing deter")
+        process_deter(
             db_url=db_url,
-            is_temp=True,
-            land_use_image=land_use_image,
-            biome=_biome,
-            force_recreate=not index,
+            biome_list=list(biome),
+            land_use_dir=Path(land_use_dir),
         )
 
-    insert_deter_in_land_use_tables(db_url=db_url, is_temp=True)
+    if ACTIVE_FIRES_INDICATOR in indicators:
+        logger.info("processing active fires")
+        process_active_fires(
+            db_url=db_url,
+            biome_list=list(biome),
+            land_use_dir=Path(land_use_dir),
+            all_data=all_data,
+        )
+
     percentage_calculation_for_areas(db_url=db_url, is_temp=True)
     reset_land_use_tables(db_url=db_url, is_temp=False)
-    copy_data_to_final_tables(db_url=db_url, all_data=all_data)
+    copy_data_to_final_tables(db_url=db_url, all_data=all_data, indicators=indicators)
 
     if drop_tmp:
-        drop_tmp_tables(db_url=db_url)
+        drop_tmp_tables(db_url=db_url, indicators=indicators)
 
 
-def _create_land_structure_table(db_url: str, table: str, force_recreate: bool):
+def create_land_structure_table(db_url: str, table: str, force_recreate: bool):
     logger.info("creating %s.", table)
     logger.debug("%s:%s", "force_recreate", force_recreate)
 
@@ -111,7 +129,7 @@ def _create_land_structure_table(db_url: str, table: str, force_recreate: bool):
             "num_pixels int4 NULL",
             "geocode varchar(80) NULL",
             "biome varchar(254) NULL",
-            "CONSTRAINT gid_landuse_biome_unique UNIQUE (gid, biome, land_use_id)",
+            "UNIQUE (gid, biome, land_use_id)",
         ],
         force_recreate=force_recreate,
     )
@@ -129,6 +147,209 @@ def _create_land_structure_table(db_url: str, table: str, force_recreate: bool):
     )
 
 
+def insert_data_in_land_use_tables(
+    data: gpd.GeoDataFrame, db: DatabaseFacade, table_prefix: str, log: bool = False
+):
+    for spatial_unit in read_spatial_units(db=db):
+        tmpspatial_unit = f"{table_prefix}{spatial_unit}"
+        logger.info("processing %s ...", tmpspatial_unit)
+
+        spatial_units = gpd.GeoDataFrame.from_postgis(
+            sql=f'SELECT suid, geometry FROM "{spatial_unit}"',
+            con=db.conn,
+            geom_col="geometry",
+        )
+
+        logger.info("joining...")
+        join = gpd.sjoin(data, spatial_units, how="inner", predicate="intersects")
+
+        logger.info("grouping...")
+        group = (
+            join[
+                [
+                    "suid",
+                    "land_use_id",
+                    "classname",
+                    "date",
+                    "num_pixels",
+                    "biome",
+                    "geocode",
+                ]
+            ]
+            .groupby(["suid", "land_use_id", "classname", "date", "biome", "geocode"])[
+                "num_pixels"
+            ]
+            .sum()
+        )
+
+        values: list = []
+        with alive_bar(len(group)) as progress_bar:
+            for key, value in group.items():
+                values.append(
+                    f"""
+                        (
+                            {key[0]},
+                            {key[1]},
+                            '{key[2]}',
+                            TIMESTAMP '{key[3].year}-{key[3].month}-{key[3].day}',
+                            {value * PIXEL_LAND_USE_AREA},
+                            '{key[4]}',
+                            '{key[5]}'
+                        )
+                    """
+                )
+                progress_bar()
+
+        assert len(values) > 0
+
+        sql = f"""
+            INSERT INTO "{tmpspatial_unit}_land_use" (
+                suid, land_use_id, classname, "date", area, geocode, biome
+            ) 
+            VALUES {','.join(values)};
+        """
+
+        logger.info("inserting data into %s_land_use", tmpspatial_unit)
+        db.execute(sql=sql, log=log)
+
+
+def process_active_fires(
+    db_url: str, biome_list: list, land_use_dir: Path, all_data: bool
+):
+    for index, biome in enumerate(biome_list):
+        logger.info("processing biome %s", biome)
+        assert is_valid_biome(biome=biome)
+
+        land_use_image = land_use_dir / f"{biome}_land_use.tif"
+        logger.debug(land_use_image)
+        assert land_use_image.exists()
+
+        process_active_fires_land_structure(
+            db_url=db_url,
+            is_temp=True,
+            land_use_image=land_use_image,
+            biome=biome,
+            all_data=all_data,
+            force_recreate=not index,
+        )
+
+    insert_fires_in_land_use_tables(db_url=db_url, is_temp=True)
+
+
+def process_active_fires_land_structure(
+    is_temp: bool,
+    land_use_image: Path,
+    db_url: str,
+    biome: str,
+    all_data: bool,
+    force_recreate: bool,
+):
+    db = DatabaseFacade.from_url(db_url=db_url)
+
+    spatial_units = read_spatial_units(db=db)
+
+    table_prefix = get_prefix(is_temp=is_temp)
+
+    table = f"{table_prefix}fires_land_structure"
+    where = f"WHERE biome='{biome}'"
+
+    if all_data:
+        create_land_structure_table(
+            db_url=db_url, table=table, force_recreate=force_recreate
+        )
+    else:
+        where += f"""
+            AND view_date > (SELECT MAX(date) FROM "{list(spatial_units.keys())[0]}_land_use" WHERE classname='{ACTIVE_FIRES_CLASSNAME}')
+        """
+
+    landuse_raster = rio.open(land_use_image)
+
+    sql = f"""
+        SELECT id as gid, biome, geocode, geom
+        FROM fires.active_fires
+        {where}
+    """
+    logger.debug(sql)
+    fires = gpd.GeoDataFrame.from_postgis(sql=sql, con=db.conn, geom_col="geom")
+    coord_list = list(zip(fires["geom"].x, fires["geom"].y))
+
+    fires["value"] = list(landuse_raster.sample(coord_list))
+    values: list = []
+
+    with alive_bar(len(fires)) as progress_bar:
+        for _, point in fires.iterrows():
+            if point["value"][0] > 0:
+                values.append(
+                    f"('{point.gid}', '{point.biome}', '{point.geocode}', {point['value'][0]}, 1)"
+                )
+            progress_bar()
+    logger.debug("len(values): %s", len(values))
+
+    assert len(values) > 0
+
+    sql = f"""
+        INSERT INTO {table_prefix}fires_land_structure (gid, biome, geocode, land_use_id, num_pixels)
+        VALUES {','.join(values)};
+    """
+
+    logger.info("inserting into %sfires_land_structure", table_prefix)
+
+    db.execute(sql=sql, log=False)
+
+
+def insert_fires_in_land_use_tables(db_url: str, is_temp: bool):
+    logger.info("Insert ACTIVE FIRES data in land use tables for each spatial units.")
+
+    db = DatabaseFacade.from_url(db_url=db_url)
+
+    table_prefix = get_prefix(is_temp=is_temp)
+
+    data = gpd.GeoDataFrame.from_postgis(
+        sql=f"""
+            SELECT
+                a.id,
+                a.land_use_id,
+                a.num_pixels,
+                a.biome,
+                a.geocode,
+                '{ACTIVE_FIRES_CLASSNAME}' as classname,
+                b.view_date AS date,
+                b.geom AS geometry
+            FROM
+                {table_prefix}fires_land_structure a 
+            INNER JOIN
+                fires.active_fires b ON a.gid = b.id::text AND a.biome = b.biome;
+        """,
+        con=db.conn,
+        geom_col="geometry",
+        crs="EPSG:4674",
+    )
+
+    insert_data_in_land_use_tables(
+        db=db, data=data, table_prefix=table_prefix, log=False
+    )
+
+
+def process_deter(db_url: str, biome_list: list, land_use_dir: Path):
+    for index, biome in enumerate(biome_list):
+        logger.info("processing biome %s", biome)
+        assert is_valid_biome(biome=biome)
+
+        land_use_image = land_use_dir / f"{biome}_land_use.tif"
+        logger.debug(land_use_image)
+        assert land_use_image.exists()
+
+        process_deter_land_structure(
+            db_url=db_url,
+            is_temp=True,
+            land_use_image=land_use_image,
+            biome=biome,
+            force_recreate=not index,
+        )
+
+    insert_deter_in_land_use_tables(db_url=db_url, is_temp=True)
+
+
 def process_deter_land_structure(
     is_temp: bool,
     land_use_image: Path,
@@ -142,7 +363,7 @@ def process_deter_land_structure(
 
     table = f"{table_prefix}deter_land_structure"
     if force_recreate:
-        _create_land_structure_table(
+        create_land_structure_table(
             db_url=db_url, table=table, force_recreate=force_recreate
         )
 
@@ -159,8 +380,9 @@ def process_deter_land_structure(
 
     landuse_raster = rio.open(land_use_image)
 
+    values: list = []
+
     with alive_bar(len(deter)) as progress_bar:
-        values: list = []
         for _, row in deter.iterrows():
             geoms = [mapping(row.geom)]
             out_image, _ = mask(landuse_raster, geoms, crop=True)
@@ -174,16 +396,17 @@ def process_deter_land_structure(
                     )
             progress_bar()
 
-        logger.debug("len(values): %s", len(values))
+    logger.debug("len(values): %s", len(values))
 
-        assert len(values) > 0
+    assert len(values) > 0
 
-        sql = f"""
-            INSERT INTO {table_prefix}deter_land_structure (gid, biome, geocode, land_use_id, num_pixels)
-            VALUES {','.join(values)};
-        """
+    sql = f"""
+        INSERT INTO {table_prefix}deter_land_structure (gid, biome, geocode, land_use_id, num_pixels)
+        VALUES {','.join(values)};
+    """
 
-        db.execute(sql=sql, log=False)
+    logger.info("inserting into %sdeter_land_structure", table_prefix)
+    db.execute(sql=sql, log=False)
 
 
 def insert_deter_in_land_use_tables(db_url: str, is_temp: bool):
@@ -193,7 +416,7 @@ def insert_deter_in_land_use_tables(db_url: str, is_temp: bool):
 
     table_prefix = get_prefix(is_temp=is_temp)
 
-    land_structure = gpd.GeoDataFrame.from_postgis(
+    data = gpd.GeoDataFrame.from_postgis(
         sql=f""" 
             SELECT 
                 a.id, 
@@ -234,82 +457,20 @@ def insert_deter_in_land_use_tables(db_url: str, is_temp: bool):
                 ) AS tb
             ) b ON a.gid = b.gid AND a.biome = b.biome
             INNER JOIN 
-                deter_class c ON b.classname = c.name
+                class c ON b.classname = c.name
             INNER JOIN 
-                deter_class_group d ON c.group_id = d.id;
+                class_group d ON c.group_id = d.id;
         """,
         con=db.conn,
         geom_col="geometry",
         crs="EPSG:4674",
     )
 
-    for spatial_unit in read_spatial_units(db=db):
-        tmpspatial_unit = f"{table_prefix}{spatial_unit}"
-        logger.info("processing %s ...", tmpspatial_unit)
-
-        spatial_units = gpd.GeoDataFrame.from_postgis(
-            sql=f'SELECT suid, geometry FROM "{spatial_unit}"',
-            con=db.conn,
-            geom_col="geometry",
-        )
-
-        logger.info("joining...")
-        join = gpd.sjoin(
-            land_structure, spatial_units, how="inner", predicate="intersects"
-        )
-
-        logger.info("grouping...")
-        group = (
-            join[
-                [
-                    "suid",
-                    "land_use_id",
-                    "classname",
-                    "date",
-                    "num_pixels",
-                    "biome",
-                    "geocode",
-                ]
-            ]
-            .groupby(["suid", "land_use_id", "classname", "date", "biome", "geocode"])[
-                "num_pixels"
-            ]
-            .sum()
-        )
-
-        with alive_bar(len(group)) as progress_bar:
-            values: list = []
-            for key, value in group.items():
-                values.append(
-                    f"""
-                        (
-                            {key[0]},
-                            {key[1]},
-                            '{key[2]}',
-                            TIMESTAMP '{key[3].year}-{key[3].month}-{key[3].day}',
-                            {value * PIXEL_LAND_USE_AREA},
-                            '{key[4]}',
-                            '{key[5]}'
-                        )
-                    """
-                )
-                progress_bar()
-
-            sql = f"""
-                INSERT INTO "{tmpspatial_unit}_land_use" (
-                    suid, land_use_id, classname, "date", area, geocode, biome
-                ) 
-                VALUES {','.join(values)};
-            """
-
-            db.execute(sql=sql, log=False)
+    insert_data_in_land_use_tables(db=db, data=data, table_prefix=table_prefix)
 
 
 def percentage_calculation_for_areas(db_url: str, is_temp: bool):
-    logger.info("Using Spatial Units areas and DETER areas to calculate percentage.")
-
-    fire_classname = "AF"
-    risk_classname = "RK"
+    logger.info("using spatial units and deter areas to calculate the percentages")
 
     db = DatabaseFacade.from_url(db_url=db_url)
 
@@ -324,16 +485,20 @@ def percentage_calculation_for_areas(db_url: str, is_temp: bool):
             WHERE
                 public."{tmpspatial_unit}_land_use".suid=su.suid
                 AND public."{tmpspatial_unit}_land_use".classname NOT IN (
-                    '{fire_classname}','{risk_classname}'
+                    '{ACTIVE_FIRES_CLASSNAME}','{RISK_CLASSNAME}'
                 )
         """
         db.execute(sql)
 
 
-def copy_data_to_final_tables(db_url: str, all_data: bool):
+def copy_data_to_final_tables(db_url: str, all_data: bool, indicators: list):
     logger.info("copying the new processed data to the final tables")
 
-    _copy_deter_land_structure(db_url=db_url, all_data=all_data)
+    if DETER_INDICATOR in indicators:
+        copy_deter_land_structure(db_url=db_url, all_data=all_data)
+
+    if ACTIVE_FIRES_INDICATOR in indicators:
+        copy_fires_land_structure(db_url=db_url, all_data=all_data)
 
     db = DatabaseFacade.from_url(db_url=db_url)
 
@@ -346,7 +511,7 @@ def copy_data_to_final_tables(db_url: str, all_data: bool):
         db.copy_table(src=tmp_land_use_table, dst=land_use_table)
 
 
-def _copy_deter_land_structure(db_url: str, all_data: bool):
+def copy_deter_land_structure(db_url: str, all_data: bool):
     table = "deter_land_structure"
 
     logger.info("copying data from %s to %s.", get_prefix(is_temp=True) + table, table)
@@ -354,7 +519,7 @@ def _copy_deter_land_structure(db_url: str, all_data: bool):
     db = DatabaseFacade.from_url(db_url=db_url)
 
     if all_data:
-        _create_land_structure_table(db_url=db_url, table=table, force_recreate=True)
+        create_land_structure_table(db_url=db_url, table=table, force_recreate=True)
 
     else:
         # here, we expect deter.tmp_data to only have DETER data coming from the current table
@@ -371,10 +536,29 @@ def _copy_deter_land_structure(db_url: str, all_data: bool):
     db.copy_table(src=f"{get_prefix(is_temp=True)}{table}", dst=table)
 
 
-def drop_tmp_tables(db_url: str):
+def copy_fires_land_structure(db_url: str, all_data: bool):
+    table = "fires_land_structure"
+
+    logger.info("copying data from %s to %s.", get_prefix(is_temp=True) + table, table)
+
+    db = DatabaseFacade.from_url(db_url=db_url)
+
+    if all_data:
+        create_land_structure_table(db_url=db_url, table=table, force_recreate=True)
+
+    # copy data from temporary table
+    db.copy_table(src=f"{get_prefix(is_temp=True)}{table}", dst=table)
+
+
+def drop_tmp_tables(db_url: str, indicators: list):
     """Drop the temporary tables."""
     db = DatabaseFacade.from_url(db_url=db_url)
     for spatial_unit in read_spatial_units(db=db):
         land_use_table = f"tmp_{spatial_unit}_land_use"
         db.drop_table(table=land_use_table)
-    db.drop_table(table="tmp_deter_land_structure")
+
+    if DETER_INDICATOR in indicators:
+        db.drop_table(table="tmp_deter_land_structure")
+
+    if ACTIVE_FIRES_INDICATOR in indicators:
+        db.drop_table(table="tmp_fires_land_structure")
